Allow users to pass options on the command line
Users can now pass their desired width as well as the path of the file
that contains text to wrap as command line arguments. The width defaults
to 80, and the path is required.

I still need to figure out how to accept text on STDIN if the path isn't
provided, but we'll get there.

There's also a bit of refactoring going on in here, but this is the real
meat of what I was trying to accomplish.

This adds optparse-applicative as a dependency (which I apparently
_accidentally_ added into the cabal file on master earlier. Oops).