Accept input on STDIN
In addition to being able to pass a file path on the commandline, it
would be nice to be able to also accept input on STDIN.

To do this, we're adding a default value of `-` to represent STDIN for
the file path argument. If the file path is set to `-`, we'll read from
STDIN. Otherwise, we'll read from the provided file path.