Allow header parsing to fail
We don't always have headers, and that's OK. If we don't we should just
go on with our day.

If you don't use `try` with `option`, then the parser always consumes
content, and so `option` never returns the default value. By using
`try`, we can ensure that if the parser fails it does _not_ consume
content and so the default can be returned.