Add simple tests
For right now, it probably makes a reasonable amount of sense to test
this by comparing static input against static output. As more things
need to be parsed, we can just add them to the input and modify them to
show how we expect them to look in the output and the tests should fail.