Fix the parsing of multi-line headers
Because `singleLine` actually swallows the `eol` already, our
`headerContinued` parser was actually matching text like the following:

```
	foo bar baz

foo bar baz
```

This wasn't intentional, and resulted in parser failures when parsing
files with multi-line headers. Instead, we should only parse to the end
of the tabbed line (which is what `singleLine` does by default)

We were also adding a newline to the end of the `headerContinued`
result, which didn't actually make sense since that is added by the
concatenation that happens when we reconstitute the continued header.
Instead, we want to add a newline to the beginning of the continued
header, since that character is being thrown away when grabbing the
previous line.

In addition, since `unlines` adds a trailing newline, using it as a part
of our parser resulted in an extra newline being added between the
headers and the body text. Using `foldl mappend ""` doesn't add this
extra newline and means that our headers stay formatted the way we
expect.