Allow code block parsing to fail
Previously, if a line started with ` and didn't signal the start of a
code block, parsing would fail but would still consume input. This would
mean that `<|>` would result in an error instead of falling back to
parsing a `Normal` line of content.

To fix this, we should use `try` to ensure that parsing `codeBlock`
content fails without consuming content.