Don't add extra newline info to pgp blocks
Previously, we weren't matching the `eol` at the end of PGP blocks. This meant
that we'd end up adding an additional newline when reconstructing the text. To
fix this, we should make sure we match on that newline.

Additionally, we can actually move _away_ from that mappend3 function and use a
more explicit blockContents function for extracting everything between and
including two sigils, while also removing the newline. I think this is a little
cleaner.