Preserve leading whitespace for Normal lines
Sometimes people use leading whitespace for formatting reasons when
composing plain text emails. It's a shame to have people go through that
trouble just to have us strip that formatting when we go reflow the
text.

To fix this, we can add an indentation level to the Normal constructor,
and take any whitespace we find into account. Then we can indent the
whole line (and all wrapped lines resulting from that line) with the
existing indentation level to maintain formatting.

Note that this breaks our existing whitespace tests (treating
whitespace-only lines as a newline). That will be fixed again in the
next commit.